// app/src/ui/pages/pcrs/__generated__/PcrItemUpdateMutation.graphql.ts
export type Acc_ProjectChangeRequest__cUpdateInput = {
  Acc_ProjectChangeRequest__c: Acc_ProjectChangeRequest__cUpdateRepresentation;
  Id: any;
};
// ...
export type PcrItemUpdateMutation$variables = {
  input: Acc_ProjectChangeRequest__cUpdateInput;
  projectId: string;
};
export type PcrItemUpdateMutation$data = {
  readonly uiapi: {
    readonly Acc_ProjectChangeRequest__cUpdate: {
      readonly success: boolean | null | undefined;
    } | null | undefined;
  };
};
export type PcrItemUpdateMutation = {
  response: PcrItemUpdateMutation$data;
  variables: PcrItemUpdateMutation$variables;
};